requirements.txt
web3
streamlit
python-dotenv